backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.22
sqlalchemy==2.0.36
alembic==1.13.2
python-multipart==0.0.9
Pillow==11.0.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
PyMuPDF==1.24.12
python-docx==1.1.2
python-pptx==1.0.2
pdf2image==1.17.0
redis==5.2.0
pydantic-settings==2.4.0
minio==7.2.7
psutil==7.1.3
supabase==2.24.0
email-validator==2.2.0
psycopg2-binary==2.9.9
PyMySQL==1.1.1
